+++ CHANGELOG.md
@@ -7,6 +7,15 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [unreleased]
 
+### Fixed
+
+* Fix edge case for keywords inside Python if/else ([#115][115])
+
+### Removed
+
+* No longer raise error if multiple keywords have the same name (e.g. `rule a` used
+  twice)
+
 ## [0.4.4]
 
 ### Fixed
@@ -224,6 +233,7 @@ is 40 character long, the line is 48 characters long. However, we were only pass
 [108]: https://github.com/snakemake/snakefmt/issues/108
 [109]: https://github.com/snakemake/snakefmt/issues/109
 [111]: https://github.com/snakemake/snakefmt/issues/111
+[115]: https://github.com/snakemake/snakefmt/issues/115
 [116]: https://github.com/snakemake/snakefmt/pull/116
 [118]: https://github.com/snakemake/snakefmt/issues/118
 [121]: https://github.com/snakemake/snakefmt/issues/121